+++ src/ostrowski/util/Diagnostics.java
@@ -460,7 +460,7 @@ public class Diagnostics {
                _prebuiltAssociations = null;
             }
             else {
-               occuranceCount = occuranceCount.intValue() + 1;
+               occuranceCount = occuranceCount + 1;
 //               duplicateAssociation = true;
             }
             valueTable.put(itemValue, occuranceCount);
@@ -483,14 +483,14 @@ public class Diagnostics {
             if (valueTable != null) {
                Integer occuranceCount = valueTable.get(itemValue);
                if (occuranceCount != null) {
-                  if (occuranceCount.intValue() <= 1) {
+                  if (occuranceCount <= 1) {
                      valueTable.remove(itemValue);
                      // clear out the pre-build association string so we force
                      // the association string to be re-built next time its used.
                      _prebuiltAssociations = null;
                   }
                   else {
-                     occuranceCount = occuranceCount.intValue() - 1;
+                     occuranceCount = occuranceCount - 1;
                      valueTable.put(itemValue, occuranceCount);
                   }
                }
